Slime on screen

+++ core/src/com/group25/game/Creature.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public abstract class Creature extends GameEntity{
 
-    public Creature(float positionX, float positionY,int width, int height,int health, Sprite img, int entitySpeed, Rectangle hitbox){
+    public Creature(float positionX, float positionY,int width, int height,int health, Sprite img, float entitySpeed, Rectangle hitbox){
         super(positionX, positionY, width, height, img, entitySpeed,hitbox);
         hitbox = new Rectangle((int)positionX,(int)positionY,width,height);
     }
