+++ core/src/com/group25/game/Enemy.java
@@ -8,7 +8,7 @@ public abstract class Enemy extends Creature{
 
     private Rectangle alertArea;
 
-    public Enemy(float positionX, float positionY, int width, int height, int health, Sprite img, int entitySpeed,Rectangle hitbox, Rectangle alertArea) {
+    public Enemy(float positionX, float positionY, int width, int height, int health, Sprite img, float entitySpeed,Rectangle hitbox, Rectangle alertArea) {
         super(positionX, positionY, width, height, health, img, entitySpeed,hitbox);
         this.alertArea = alertArea;
     }
